further conversions to modules
